+++ MG Engine/src/Math/Vector4.h
@@ -48,7 +48,7 @@ public:
 	static float Dot(Vector4&& X, Vector4&& Y);
 
 	std::string ToString() const override;
-	int GetHashCode() const override;
+	size_t GetHashCode() const override;
 
 	~Vector4() { }
 };
